remove bind_column_references method and its last usages
